Reorganize server structure
