+++ .gitignore
@@ -29,7 +29,7 @@ dist/
 node_modules/
 bower_components/
 client/lib
-server/env/config.js
+server/config/env/config.js
 
 
 # Floobits
