+++ server/config/middleware.js
@@ -1,0 +1,9 @@
+var bodyParser = require('body-parser');
+var morgan = require('morgan');
+
+module.exports = function(app, express) {
+	app.use(morgan('dev'));
+	app.use(bodyParser.json());
+	// app.use('/', express.static('../client'));
+	app.use(express.static('../client'));
+};
